back_app/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.1
tzdata==2024.2
